requirements.txt
python-pptx>=1.0.2
markdown-it-py>=4.0.0
mdit-py-plugins>=0.4.0
click>=8.3.0
PyYAML>=6.0.3
google-generativeai>=0.8.0
pydantic>=2.0.0
